+++ test/test_helper.rb
@@ -42,6 +42,10 @@ def elasticsearch_below60?
   Searchkick.server_below?("6.0.0-alpha1")
 end
 
+def elasticsearch_below61?
+  Searchkick.server_below?("6.1.0-alpha1")
+end
+
 def nobrainer?
   defined?(NoBrainer)
 end
@@ -409,7 +413,6 @@ class Product
     word_middle: [:name],
     word_end: [:name],
     highlight: [:name],
-    searchable: [:name, :color],
     filterable: [:name, :color, :description],
     similarity: "BM25",
     match: ENV["MATCH"] ? ENV["MATCH"].to_sym : nil
@@ -440,7 +443,6 @@ end
 
 class Store
   searchkick \
-    default_fields: elasticsearch_below60? ? nil : [:name],
     routing: true,
     merge_mappings: true,
     mappings: {
@@ -462,7 +464,6 @@ end
 
 class Region
   searchkick \
-    default_fields: elasticsearch_below60? ? nil : [:name],
     geo_shape: {
       territory: {tree: "quadtree", precision: "10km"}
     }
@@ -480,7 +481,6 @@ end
 
 class Speaker
   searchkick \
-    default_fields: elasticsearch_below60? ? nil : [:name],
     conversions: ["conversions_a", "conversions_b"]
 
   attr_accessor :conversions_a, :conversions_b, :aisle
@@ -496,7 +496,6 @@ end
 
 class Animal
   searchkick \
-    default_fields: elasticsearch_below60? ? nil : [:name],
     inheritance: !elasticsearch_below60?,
     text_start: [:name],
     suggest: [:name],